+++ docs/source/conf.py
@@ -13,7 +13,7 @@ sys.path.append(os.path.abspath('../..'))
 project = 'Metratec rfid'
 copyright = '2023, Metratec GmbH'
 author = 'www.metratec.com'
-release = '1.1.0'
+release = '1.1.1'
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration