+++ CHANGELOG.md
@@ -1,11 +1,17 @@
 # Release Notes
 
+## 1.1.1
+
+* reader type check updated
+* bug fixes
+
 ## 1.1.0
 
 * Gen2 UHF
   * reader GRU300 and QRG2 added
   * special impinj tag methods added
   * bug fixes
+* send_custom_command method added
 
 ## 1.0.0